requirements_gpu.txt
onnx==1.16.1
onnxruntime-gpu==1.20.1
numpy==1.26.3
pyclipper==1.3.0.post5
shapely==2.0.3
opencv-python==4.11.0.86
pillow==11.0.0